finish journal + link front-back end of checklist

+++ server/src/resolvers/folders.ts
@@ -1,13 +1,14 @@
 import { Folder } from "../entities/Folder";
 import { MyContext } from "../types";
-import { Arg, Ctx, Field, InputType, Int, Mutation, Query, Resolver, UseMiddleware } from "type-graphql";
+import { Arg, Ctx, Field, InputType, Mutation, Query, Resolver, UseMiddleware } from "type-graphql";
 import { isAuth } from "../middleware/isAuth";
-import { IsNull } from "typeorm";
+import { getRepository, IsNull } from "typeorm";
+import { Entry } from "../entities/Entry";
 
 @InputType()
 class FolderOptionsInput {
-    @Field(() => Int, { nullable: true })
-    folderId: number;
+    @Field({ nullable: true })
+    folderId: string;
     @Field({ nullable: true })
     title: string;
 }
@@ -16,35 +17,35 @@ class FolderOptionsInput {
 export class FolderResolver {
     @Query(() => [Folder])
     folders(): Promise<Folder[]> {
-        return Folder.find({});
+        return Folder.find({ order: { updatedAt: "DESC" } });
     }
 
     @Query(() => Folder)
     folder(
-        @Arg("id", () => Int) id: number
+        @Arg("id") id: string
     ): Promise<Folder | undefined> {
         return Folder.findOne({ id });
     }
 
     @Query(() => [Folder])
     @UseMiddleware(isAuth)
-    myFolders(
+    async myFolders(
         @Ctx() { req }: MyContext,
-        @Arg("rootFolderId", () => Int, { nullable: true }) rootFolderId: number
+        @Arg("rootFolderId", { nullable: true }) rootFolderId: string
     ): Promise<Folder[]> {
         if (rootFolderId) {
-            return Folder.find({ userId: req.session.userId, rootFolderId })
+            return await Folder.find({ order: { updatedAt: "DESC" }, where: { userId: req.session.userId, rootFolderId } })
         }
-        return Folder.find({ userId: req.session.userId, rootFolderId: IsNull() })
+        return await Folder.find({ order: { updatedAt: "DESC" }, where: { userId: req.session.userId, rootFolderId: IsNull() } })
     }
 
     @Query(() => Folder)
     @UseMiddleware(isAuth)
     myFolder(
         @Ctx() { req }: MyContext,
-        @Arg("id", () => Int) id: number
+        @Arg("id") id: string
     ): Promise<Folder | undefined> {
-        return Folder.findOne({ userId: req.session.userId, id })
+        return Folder.findOne({ where: { userId: req.session.userId, id }, relations: ["content", "children"] })
     }
 
     @Mutation(() => Folder)
@@ -63,19 +64,62 @@ export class FolderResolver {
         return folder;
     }
 
-    @Mutation(() => Boolean)
+    @Mutation(() => String, { nullable: true })
     @UseMiddleware(isAuth)
     async deleteFolder(
         @Ctx() { req }: MyContext,
-        @Arg("id", () => Int) id: number,
-    ): Promise<Boolean> {
+        @Arg("id") id: string
+    ): Promise<string | null> {
+        const folder = await Folder.findOne({ id })
+
+        if (!folder) {
+            return null;
+        }
+
         try {
+            // Redirect all items within the folder to its root folder to prevent unintended item loss
+            getRepository(Entry).createQueryBuilder()
+                .update()
+                .set({ rootFolderId: folder.rootFolderId })
+                .where(`rootFolderId = :rootFolderId`, { rootFolderId: id })
+                .execute();
+            getRepository(Folder).createQueryBuilder()
+                .update()
+                .set({ rootFolderId: folder.rootFolderId })
+                .where(`rootFolderId = :rootFolderId`, { rootFolderId: id })
+                .execute();
+
             Folder.delete({ id, userId: req.session.userId })
         }
         catch {
-            return false;
+            return null;
         }
 
-        return true;
+        return id;
+    }
+
+    @Mutation(() => Folder)
+    @UseMiddleware(isAuth)
+    async updateFolder(
+        @Ctx() { req }: MyContext,
+        @Arg("id") id: string,
+        @Arg("options") options: FolderOptionsInput
+    ): Promise<Folder | undefined> {
+
+        const folder = await Folder.findOne({ id, userId: req.session.userId })
+
+        if (!folder) {
+            return undefined;
+        }
+
+        if (typeof options.title !== "undefined") {
+            Folder.update({ id }, { title: options.title })
+        }
+
+        if (typeof options.folderId !== "undefined") {
+            Folder.update({ id }, { rootFolderId: options.folderId })
+        }
+
+        return await Folder.findOne({ id, userId: req.session.userId });
     }
 }
